backend/requirements.txt
asgiref==3.6.0
Django==4.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.3